// packages/pqb/src/query/extra-features/get-column-info/get-column-info.d.ts
import { Column } from '../../../columns';
import { PickQueryShape } from '../../pick-query-types';
import { QueryThenShallowSimplify } from '../../then/then';
/**
 * Result type for `columnInfo` method.
 * Sets query kind to 'columnInfo', returns a single value (may return undefined),
 * the value is a {@link GetColumnInfo} object or a Record with keys for column names and ColumnInfo objects as values.
 **/
export type SetQueryReturnsColumnInfo<T extends PickQueryShape, Column extends keyof T['shape'] | undefined, Result = Column extends keyof T['shape'] ? GetColumnInfo : {
    [K in keyof T['shape']]: GetColumnInfo;
}> = Omit<T, 'result' | 'returnType' | 'then'> & {
    result: {
        value: Column.Pick.QueryColumnOfType<Result>;
    };
    returnType: 'value';
    then: QueryThenShallowSimplify<Result>;
};
export type GetColumnInfo = {
    defaultValue: unknown;
    type: string;
    maxLength: number | null;
    nullable: boolean;
};
/**
 * Returns an object with the column info about the current table, or an individual column if one is passed, returning an object with the following keys:
 *
 * ```ts
 * type ColumnInfo = {
 *   defaultValue: unknown; // the default value for the column
 *   type: string; // the column type
 *   maxLength: number | null; // the max length set for the column, present on string types
 *   nullable: boolean; // whether the column may be null
 * };
 *
 * import { getColumnInfo } from 'orchid-orm';
 *
 * // columnInfo has type Record<string, ColumnInfo>, where string is name of columns
 * const columnInfo = await getColumnInfo(db.table);
 *
 * // singleColumnInfo has the type ColumnInfo
 * const singleColumnInfo = await getColumnInfo(db.table, 'name');
 * ```
 *
 * @param column - optional: select info for only a single column if provided, or for all table columns if not
 */
export declare function getColumnInfo<T extends PickQueryShape, Column extends keyof T['shape'] | undefined = undefined>(query: T, column?: Column): SetQueryReturnsColumnInfo<T, Column>;